Refactor names and exports
- Refactor all names for all helper functions.
- Refactor all export to default for the helper functions.
