+++ src/helpers/sold-television.js
@@ -1,9 +1,11 @@
 import {inventory} from '../constants/inventory.js';
 
-export function soldTelivision() {
+function soldTelevision() {
   let totalSold = 0;
   for (let i = 0; i < inventory.length; i++) {
     totalSold += inventory[i].sold;
   }
   return totalSold;
 }
+
+export default soldTelevision;
