--- src/helpers/tvtobesold.js
@@ -1,6 +1,0 @@
-import {soldTelivision} from "./soldTelivision.js";
-import {totalPurchases} from "./totalPurchases.js";
-
-export function tvtobesold() {
-  return totalPurchases() - soldTelivision();
-}
